requirements.txt
# Flask==2.2.3
# SQLAlchemy~=2.0.4
elasticsearch~=8.6.2
# Flask-Migrate==2.5.3
# flask-script==2.0.5
# flask_marshmallow
pandas~=1.5.3
# psycopg2-binary==2.9.3
# Flask-SQLAlchemy==2.5.1
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.27
Flask-Script==2.0.6
Flask-Migrate==2.5.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
Werkzeug==2.0.3
Jinja2==3.0.0
itsdangerous==1.1.0